requirements.txt
adb_shell==0.4.4
pyserial==3.5
